+++ surface/src/simplification_remove_unused_vertices.cpp
@@ -45,26 +45,26 @@
 void
 pcl::surface::SimplificationRemoveUnusedVertices::simplify(const pcl::PolygonMesh& input, pcl::PolygonMesh& output, std::vector<int>& indices)
 {
-  if ( input.polygons.size() == 0)
+  if (input.polygons.size () == 0)
     return;
 
   unsigned int nr_points = input.cloud.width * input.cloud.height;
 
-  std::vector<int> new_indices(nr_points, -1);
-  indices.clear();
-  indices.reserve(nr_points);
+  std::vector<int> new_indices (nr_points, -1);
+  indices.clear ();
+  indices.reserve (nr_points);
 
   // mark all points in triangles as being used
-  for ( size_t polygon = 0; polygon < input.polygons.size(); ++polygon)
-    for ( size_t point = 0; point < input.polygons[polygon].vertices.size(); ++point )
-      if ( new_indices[ input.polygons[polygon].vertices[point] ] == -1 )
+  for (size_t polygon = 0; polygon < input.polygons.size (); ++polygon)
+    for (size_t point = 0; point < input.polygons[polygon].vertices.size (); ++point)
+      if (new_indices[ input.polygons[polygon].vertices[point] ] == -1 )
       {
-        new_indices[ input.polygons[polygon].vertices[point] ] = indices.size();
-        indices.push_back(input.polygons[polygon].vertices[point]);
+        new_indices[input.polygons[polygon].vertices[point]] = static_cast<int> (indices.size ());
+        indices.push_back (input.polygons[polygon].vertices[point]);
       }
 
   // in case all points are used , do nothing and return input mesh
-  if ( indices.size() == nr_points )
+  if (indices.size () == nr_points)
   {
     output = input;
     return;
@@ -72,31 +72,32 @@ pcl::surface::SimplificationRemoveUnusedVertices::simplify(const pcl::PolygonMes
 
   // copy cloud information
   output.header = input.header;
-  output.cloud.data.clear();
+  output.cloud.data.clear ();
   output.cloud.header = input.cloud.header;
   output.cloud.fields = input.cloud.fields;
   output.cloud.row_step = input.cloud.row_step;
   output.cloud.point_step = input.cloud.point_step;
   output.cloud.is_bigendian = input.cloud.is_bigendian;
   output.cloud.height = 1; // cloud is no longer organized
-  output.cloud.width = indices.size();
+  output.cloud.width = static_cast<int> (indices.size ());
   output.cloud.row_step = output.cloud.point_step * output.cloud.width;
   output.cloud.data.resize (output.cloud.width * output.cloud.height * output.cloud.point_step);
   output.cloud.is_dense = false;
-  output.polygons.clear();
+  output.polygons.clear ();
 
   // copy (only!) used points
-  for ( size_t i = 0; i < indices.size(); ++i )
+  for (size_t i = 0; i < indices.size (); ++i)
     memcpy (&output.cloud.data[i * output.cloud.point_step], &input.cloud.data[indices[i] * output.cloud.point_step], output.cloud.point_step);
 
   // copy mesh information (and update indices)
-  output.polygons.reserve( input.polygons.size() );
-  for ( size_t polygon = 0; polygon < input.polygons.size(); ++polygon)
+  output.polygons.reserve (input.polygons.size ());
+  for (size_t polygon = 0; polygon < input.polygons.size (); ++polygon)
   {
     pcl::Vertices corrected_polygon;
-    corrected_polygon.vertices.resize(input.polygons[polygon].vertices.size());
-    for ( size_t point = 0; point < input.polygons[polygon].vertices.size(); ++point )
-      corrected_polygon.vertices[point] = new_indices[ input.polygons[polygon].vertices[point] ];
-    output.polygons.push_back(corrected_polygon);
+    corrected_polygon.vertices.resize (input.polygons[polygon].vertices.size ());
+    for (size_t point = 0; point < input.polygons[polygon].vertices.size(); ++point)
+      corrected_polygon.vertices[point] = new_indices[input.polygons[polygon].vertices[point]];
+    output.polygons.push_back (corrected_polygon);
   }
 }
+
